+++ kg_hide_menu/__manifest__.py
@@ -17,7 +17,7 @@
     'website': "https://www.klystronglobal.com/",
     'images': ["static/description/banner.png"],
     'category': 'Extra Rights',
-    'version': "14.0.1.0.0",
+    'version': "16.0.1.0.0",
     'license': 'AGPL-3',
     'depends': [
         'base'