+++ kg_hide_menu/models/res_users.py
@@ -14,3 +14,8 @@ class ResUsers(models.Model):
 
     hide_menu_access_ids = fields.Many2many('ir.ui.menu', 'ir_ui_hide_menu_rel', 'uid', 'menu_id',
                                             string='Hide Access Menu')
+
+    def write(self, vals):
+        res = super(ResUsers, self).write(vals)
+        self.self.clear_caches()
+        return res