+++ low_stocks_product_alert/doc/RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 ## Module <low_stocks_product_alert>
 
-#### 06.06.2022
-#### Version 14.0.1.0.0
+#### 27.07.2023
+#### Version 16.0.1.0.0
 #### ADD
-- Initial commit for Product Stock Alert Module
+- Initial commit for Product Low Stock Alert
+
+#### 10.06.2024
+#### Version 16.0.1.0.1
+##### BUGFIX
+- Fixed the error in displaying the alert quantity for individual products